fix(Windows): Incorrect pointer cast
The `AddressOfFuntions` array are actually 32-bit RVAs, not 16-bit.

The cast should be done to `uint` instead of `ushort`.
This bug is not present in the abstractions branch.

Fixes #36

+++ MemoryModule/Windows/NativeAssemblyImpl.cs
@@ -661,7 +661,7 @@ namespace MemoryModule.Windows
             }
 
             // AddressOfFunctions contains the RVAs to the "real" functions
-            return (void**)(codeBase + (*((ushort*)(codeBase + exports->AddressOfFunctions + (idx * 4)))));
+            return (void**)(codeBase + (*((uint*)(codeBase + exports->AddressOfFunctions + (idx * 4)))));
         }
 #endregion
